+++ imagesgallery/__init__.py
@@ -4,4 +4,4 @@ Init for the ImagesGalleryXBlock package.
 
 from .imagesgallery import ImagesGalleryXBlock
 
-__version__ = '0.1.0'
+__version__ = '0.2.0'