+++ CHANGELOG.rst
@@ -23,3 +23,19 @@ Added
 =====
 
 * First release on PyPI.
+
+0.2.0 - 2023-09-06
+**********************************************
+
+Added
+=====
+
+* Introduced the `react-dropzone` library for drag-and-drop image file uploads.
+* Added the `react-images-gallery` library for image carousel functionality.
+
+
+Changed
+=======
+
+* Modified the XBlock to allow users to select and preview images in the dropzone.
+* Enhanced the XBlock to enable users to view the images gallery.